Fail on log
Currently, when all is well, nothing should be logged at all. We can
leverage this to use the CLI tool as a linter, and create CI jobs with
it.
Here, a new option called --fail-on-log is introduced, and it will use a
failure exit code as soon as any log is emitted.
This is based on the assumption that the logger is never and will never
be called for anything that should not result in a change in the
documentation being processed.