[!!!TASK] Make all classes final
This library has a very large number of extension points via
interfaces and abstract classes. All other classes should not
be extended by users and shall theirfor be marked as final.

There are a few classes left that should also be marked as final
but those require some extra attention.